+++ src/app/quiz/components/Carousel/index.tsx
