+++ src/app/quiz/components/Comment/CommentList.tsx
@@ -4,9 +4,8 @@ import clsx from 'clsx';
 import Image from 'next/image';
 import { Children, HTMLAttributes, useState } from 'react';
 
+import { Text } from '@/common';
 import { ICON_URL } from '@/common/constants';
-
-import { Text } from '../../../shared';
 
 interface CommentListProps {
   children: React.ReactNode;
