+++ src/app/quiz/components/Comment/LikeButton.tsx
@@ -4,9 +4,8 @@ import clsx from 'clsx';
 import Image from 'next/image';
 import { HTMLAttributes, useState } from 'react';
 
+import { Text } from '@/common';
 import { ICON_URL } from '@/common/constants';
-
-import { Text } from '../../../shared';
 
 interface LikeButtonProps
   extends Omit<HTMLAttributes<HTMLSpanElement>, 'className'> {
